extensions/discord-voice/sidecar/pyproject.toml
[project]
name = "openclaw-voice-sidecar"
version = "0.1.0"
description = "Audio sidecar for OpenClaw Discord voice agent (Estonian STT/TTS)"
requires-python = ">=3.11"
dependencies = [
    "discord.py[voice]>=2.4",
    "aiohttp>=3.9",
]

[project.optional-dependencies]
sherpa = ["sherpa-onnx>=1.12"]

[project.scripts]
openclaw-voice = "openclaw_voice.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/openclaw_voice"]
